requirements_v2.txt
# Infinite AI Security V2 - Modern dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
websockets==13.1
python-multipart==0.0.12
aiofiles==24.1.0
aiohttp==3.13.2
requests==2.32.3
numpy==2.2.0
pandas==2.2.3
prometheus-client==0.21.1
pytest==8.4.2
pytest-asyncio==1.2.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.28.1